Create Spending Tracking page interface and reconfigured Transactions and Spend Tracker page to use it..
