+++ budget-planner/src/specific-components/spend-tracking/spend-tracking-graphs.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { InnerSpendTrackingComponentProps } from ".";
+
+export class SpendTrackingGraph extends React.Component<InnerSpendTrackingComponentProps> {
+    render(): React.ReactNode {
+        return (
+            <p>Spend Tracking Graph</p>
+        )
+    }
+}
