+++ budget-planner/src/ts-utils/react-types.ts
@@ -1,0 +1,1 @@
+export type PropsOnlyComponent<P = {}> = React.Component<P> | React.FunctionComponent<P>
